# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DecisionForge API
  description: Registre immuable de décisions, recommandations déterministes, analyses
    bornées de sensibilité, comparaisons, consensus, calibration et stabilité chronologique
    recalculés à partir des preuves persistées, avec attribution chronologique des
    changements. La couverture des critères est également qualifiée sur des séries
    compatibles.
  version: 1.0.7
paths:
...
  /v1/decisions:
    get:
      tags:
      - decisions
      summary: List Decisions
      operationId: list_decisions_v1_decisions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DecisionSnapshot'
                type: array
                title: Response List Decisions V1 Decisions Get
    post:
      tags:
      - decisions
      summary: Create Decision
      operationId: create_decision_v1_decisions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionSnapshot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decisions/{decision_id}:
    get:
      tags:
      - decisions
      summary: Get Decision
      operationId: get_decision_v1_decisions__decision_id__get
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionSnapshot'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evaluations:
    post:
      tags:
      - evaluations
      summary: Create Evaluation
      operationId: create_evaluation_v1_evaluations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evaluations/{evaluation_id}:
    get:
      tags:
      - evaluations
      summary: Get Evaluation
      operationId: get_evaluation_v1_evaluations__evaluation_id__get
      parameters:
      - name: evaluation_id
        in: path
        required: true
        schema:
          type: string
          title: Evaluation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sensitivity-analyses:
    post:
      tags:
      - sensitivity
      summary: Create Sensitivity Analysis
      operationId: create_sensitivity_analysis_v1_sensitivity_analyses_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensitivityAnalysisCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensitivityAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sensitivity-analyses/{analysis_id}:
    get:
      tags:
      - sensitivity
      summary: Get Sensitivity Analysis
      operationId: get_sensitivity_analysis_v1_sensitivity_analyses__analysis_id__get
      parameters:
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
          title: Analysis Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensitivityAnalysis'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decision-comparisons:
    post:
      tags:
      - comparisons
      summary: Create Decision Comparison
      operationId: create_decision_comparison_v1_decision_comparisons_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionComparisonCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decision-comparisons/{comparison_id}:
    get:
      tags:
      - comparisons
      summary: Get Decision Comparison
      operationId: get_decision_comparison_v1_decision_comparisons__comparison_id__get
      parameters:
      - name: comparison_id
        in: path
        required: true
        schema:
          type: string
          title: Comparison Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/consensus-dossiers:
    post:
      tags:
      - consensus
      summary: Create Consensus Dossier
      operationId: create_consensus_dossier_v1_consensus_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsensusDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsensusDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/consensus-dossiers/{dossier_id}:
    get:
      tags:
      - consensus
      summary: Get Consensus Dossier
      operationId: get_consensus_dossier_v1_consensus_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsensusDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/calibration-dossiers:
    get:
      tags:
      - calibration
      summary: List Calibration Dossiers
      operationId: list_calibration_dossiers_v1_calibration_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CalibrationDossier'
                type: array
                title: Response List Calibration Dossiers V1 Calibration Dossiers
                  Get
    post:
      tags:
      - calibration
      summary: Create Calibration Dossier
      operationId: create_calibration_dossier_v1_calibration_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalibrationDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrationDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/calibration-dossiers/{dossier_id}:
    get:
      tags:
      - calibration
      summary: Get Calibration Dossier
      operationId: get_calibration_dossier_v1_calibration_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalibrationDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decision-stability-dossiers:
    get:
      tags:
      - stability
      summary: List Decision Stability Dossiers
      operationId: list_decision_stability_dossiers_v1_decision_stability_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DecisionStabilityDossier'
                type: array
                title: Response List Decision Stability Dossiers V1 Decision Stability
                  Dossiers Get
    post:
      tags:
      - stability
      summary: Create Decision Stability Dossier
      operationId: create_decision_stability_dossier_v1_decision_stability_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionStabilityDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionStabilityDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decision-stability-dossiers/{dossier_id}:
    get:
      tags:
      - stability
      summary: Get Decision Stability Dossier
      operationId: get_decision_stability_dossier_v1_decision_stability_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionStabilityDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decision-change-attribution-dossiers:
    get:
      tags:
      - attribution
      summary: List Decision Change Attribution Dossiers
      operationId: list_decision_change_attribution_dossiers_v1_decision_change_attribution_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DecisionChangeAttributionDossier'
                type: array
                title: Response List Decision Change Attribution Dossiers V1 Decision
                  Change Attribution Dossiers Get
    post:
      tags:
      - attribution
      summary: Create Decision Change Attribution Dossier
      operationId: create_decision_change_attribution_dossier_v1_decision_change_attribution_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionChangeAttributionDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionChangeAttributionDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/decision-change-attribution-dossiers/{dossier_id}:
    get:
      tags:
      - attribution
      summary: Get Decision Change Attribution Dossier
      operationId: get_decision_change_attribution_dossier_v1_decision_change_attribution_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionChangeAttributionDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/criterion-coverage-dossiers:
    get:
      tags:
      - coverage
      summary: List Criterion Coverage Dossiers
      operationId: list_criterion_coverage_dossiers_v1_criterion_coverage_dossiers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CriterionCoverageDossier'
                type: array
                title: Response List Criterion Coverage Dossiers V1 Criterion Coverage
                  Dossiers Get
    post:
      tags:
      - coverage
      summary: Create Criterion Coverage Dossier
      operationId: create_criterion_coverage_dossier_v1_criterion_coverage_dossiers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionCoverageDossierCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionCoverageDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/criterion-coverage-dossiers/{dossier_id}:
    get:
      tags:
      - coverage
      summary: Get Criterion Coverage Dossier
      operationId: get_criterion_coverage_dossier_v1_criterion_coverage_dossiers__dossier_id__get
      parameters:
      - name: dossier_id
        in: path
        required: true
        schema:
          type: string
          title: Dossier Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionCoverageDossier'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConsensusDossierCreate:
      properties:
        evaluation_ids:
          items:
            type: string
          type: array
          maxItems: 50
          minItems: 3
          title: Evaluation Ids
      additionalProperties: false
      type: object
      required:
      - evaluation_ids
      title: ConsensusDossierCreate
...
    DecisionComparisonCreate:
      properties:
        baseline_evaluation_id:
          type: string
          minLength: 1
          title: Baseline Evaluation Id
        candidate_evaluation_id:
          type: string
          minLength: 1
          title: Candidate Evaluation Id
      additionalProperties: false
      type: object
      required:
      - baseline_evaluation_id
      - candidate_evaluation_id
      title: DecisionComparisonCreate
    DecisionCreate:
      properties:
        title:
          type: string
          maxLength: 300
          minLength: 1
          title: Title
        problem:
          type: string
          maxLength: 5000
          minLength: 1
          title: Problem
        context:
          type: string
          maxLength: 5000
          title: Context
          default: ''
        alternatives:
          items:
            $ref: '#/components/schemas/AlternativeCreate'
          type: array
          maxItems: 20
          minItems: 2
          title: Alternatives
        criteria:
          items:
            $ref: '#/components/schemas/CriterionCreate'
          type: array
          maxItems: 20
          minItems: 1
          title: Criteria
        observations:
          items:
            $ref: '#/components/schemas/ObservationCreate'
          type: array
          maxItems: 400
          title: Observations
        forecast_outcome:
          anyOf:
          - $ref: '#/components/schemas/ForecastOutcomeCreate'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - title
      - problem
      - alternatives
      - criteria
      title: DecisionCreate
    DecisionSnapshot:
      properties:
        id:
          type: string
          title: Id
        input_hash:
          type: string
          title: Input Hash
        specification:
          $ref: '#/components/schemas/DecisionCreate'
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - input_hash
      - specification
      - created_at
      title: DecisionSnapshot
...
    Evaluation:
      properties:
        id:
          type: string
          title: Id
        decision_id:
          type: string
          title: Decision Id
        method:
          type: string
          title: Method
        status:
          type: string
          enum:
          - RECOMMENDED
          - BLOCKED
          - INSUFFICIENT
          title: Status
        recommended_alternative_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Recommended Alternative Key
        ranking:
          items:
            $ref: '#/components/schemas/RankedAlternative'
          type: array
          title: Ranking
        explanations:
          items:
            type: string
          type: array
          title: Explanations
        decision_hash:
          type: string
          title: Decision Hash
        outcome_hash:
          type: string
          title: Outcome Hash
        reproducible:
          type: boolean
          title: Reproducible
          default: true
        idempotent_replay:
          type: boolean
          title: Idempotent Replay
          default: false
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
        warning:
          type: string
          title: Warning
          default: Recommandation calculée à partir des données fournies ; elle ne
            prouve ni leur exactitude ni l'absence de risques externes. Une validation
            humaine reste nécessaire.
      type: object
      required:
      - id
      - decision_id
      - method
      - status
      - recommended_alternative_key
      - ranking
      - explanations
      - decision_hash
      - outcome_hash
      - created_at
      title: Evaluation
    EvaluationCreate:
      properties:
        decision_id:
          type: string
          minLength: 1
          title: Decision Id
        method:
          type: string
          const: weighted-evidence-v1
          title: Method
          default: weighted-evidence-v1
      additionalProperties: false
      type: object
      required:
      - decision_id
      title: EvaluationCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SensitivityAnalysis:
      properties:
        id:
          type: string
          title: Id
        evaluation_id:
          type: string
          title: Evaluation Id
        decision_id:
          type: string
          title: Decision Id
        method:
          type: string
          title: Method
        policy:
          type: string
          title: Policy
        qualification:
          type: string
          enum:
          - ROBUST
          - FRAGILE
          - INSUFFICIENT
          title: Qualification
        baseline_winner_alternative_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Baseline Winner Alternative Key
        winner_stability:
          type: number
          title: Winner Stability
        minimum_winner_margin:
          anyOf:
          - type: number
          - type: 'null'
          title: Minimum Winner Margin
        scenarios:
          items:
            $ref: '#/components/schemas/SensitivityScenario'
          type: array
          title: Scenarios
        reasons:
          items:
            type: string
          type: array
          title: Reasons
        evaluation_hash:
          type: string
          title: Evaluation Hash
        analysis_hash:
          type: string
          title: Analysis Hash
        reproducible:
          type: boolean
          title: Reproducible
          default: true
        idempotent_replay:
          type: boolean
          title: Idempotent Replay
          default: false
        immutable:
          type: boolean
          title: Immutable
          default: true
        created_at:
          type: string
          title: Created At
        warning:
          type: string
          title: Warning
          default: Cette analyse de sensibilité décrit uniquement la stabilité sous
            la politique de perturbation indiquée. Elle ne déclenche aucune action
            et ne remplace pas une validation humaine.
      additionalProperties: false
      type: object
      required:
      - id
      - evaluation_id
      - decision_id
      - method
      - policy
      - qualification
      - baseline_winner_alternative_key
      - winner_stability
      - minimum_winner_margin
      - scenarios
      - reasons
      - evaluation_hash
      - analysis_hash
      - created_at
      title: SensitivityAnalysis
    SensitivityAnalysisCreate:
      properties:
        evaluation_id:
          type: string
          minLength: 1
          title: Evaluation Id
        method:
          type: string
          const: weight-sensitivity-v1
          title: Method
          default: weight-sensitivity-v1
        policy:
          type: string
          const: bounded-oat-10pct-v1
          title: Policy
          default: bounded-oat-10pct-v1
      additionalProperties: false
      type: object
      required:
      - evaluation_id
      title: SensitivityAnalysisCreate
...